[fix] json解析

+++ src/main/java/com/astra/actionconfig/config/data/FixedArea.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public class FixedArea {
     public String id = "";
     public String content = "";
-    public int heightToWidthRatio = 1;
+    public double heightToWidthRatio = 1;
     public float width = 0.1f;
     public Point2F[] area    = new Point2F[4];
     public Point2F center;
